Update DistanceTraveledSinceCodesClearedObdCommand.java
Java returns an error because KM is an integer, not a float

+++ src/main/java/pt/lighthouselabs/obd/commands/control/DistanceTraveledSinceCodesClearedObdCommand.java
@@ -49,7 +49,7 @@ public class DistanceTraveledSinceCodesClearedObdCommand extends ObdCommand
 
   @Override
   public String getFormattedResult() {
-    return String.format("%.2f%s", km, "km");
+    return String.format("%.2f%s", (float)km, "km");
   }
 
   @Override
